pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "fintools"
version = "0.1.0"
description = "Shared utilities for empirical asset pricing research"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5",
    "psycopg2-binary>=2.9",
    "numpy>=1.21",
    "pyarrow>=10.0",
    "matplotlib>=3.5",
    "statsmodels>=0.14",
    "scipy>=1.9",
    "numba>=0.57",
]
